Add bitwise not support
xcc optimises constant not uses but at one point calls
wrap_value(.., type_size(..), ..) expecting the size
to be the number of 8 bit blocks but it actually gives
the number of byte sized blocks on the target.
On xtoy bytes and ints are 16 bit so sizeof(int) = 1
not 4 as wrap_value was expecting.